Added owner command tests.
